jarvis-backend/pyproject.toml
[project]
name = "jarvis-backend"
version = "0.1.0"
description = "Jarvis Backend in Python"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.55.0",
    "fastapi>=0.115.14",
    "openai>=1.93.0",
    "openai-agents>=0.1.0",
    "pydantic>=2.11.7",
    "pyright>=1.1.402",
    "python-dotenv>=1.1.1",
    "pytz>=2024.2",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=8.4.1",
]
